GEODE-9758: Move ClassUtils to geode-common
Remove "JUnit" from the name of a couple tests.
