+++ geode-serialization/src/test/java/org/apache/geode/internal/serialization/SerializationDependenciesTest.java
@@ -43,6 +43,7 @@ public class SerializationDependenciesTest {
           resideInAPackage("org.apache.geode.internal.serialization..")
               .or(not(resideInAPackage("org.apache.geode..")))
               .or(resideInAPackage("org.apache.geode.annotations.."))
-              .or(resideInAPackage("org.apache.geode.logging.."))
+              .or(resideInAPackage("org.apache.geode.internal.lang.utils.."))
+              .or(resideInAPackage("org.apache.geode.logging.internal.log4j.api.."))
               .or(resideInAPackage("org.apache.geode.test..")));
 }
